more clear language in account settings
