+++ client/components/Account/AccountPage.vue
@@ -6,21 +6,7 @@
     <hr/>
     <section>
       <AddCredentialForm />
-      <fieldset>
-        <legend>Users with access to shared account</legend>
-        <ul>
-          <li v-for="username in $store.state.account ? $store.state.account.credentials : []"
-              :key="username">
-            <button class="btn-secondary"
-                    @click="() => deleteCredential(username)">
-              Revoke Access
-            </button>
-            <span>
-              @{{username}}
-            </span>
-          </li>
-        </ul>
-      </fieldset>
+      <AccessList />
     </section>
     <hr/>
     <section>
@@ -33,6 +19,7 @@
 <script>
 import ChangeAccountNameForm from '@/components/Account/ChangeAccountNameForm.vue';
 import AddCredentialForm from '@/components/Account/AddCredentialForm.vue';
+import AccessList from '@/components/Account/AccessList.vue';
 import DeleteAccountForm from '@/components/Account/DeleteAccountForm.vue';
 
 export default {
@@ -40,6 +27,7 @@ export default {
   components: {
     ChangeAccountNameForm,
     AddCredentialForm,
+    AccessList,
     DeleteAccountForm,
   },
   mounted() {
@@ -52,34 +40,5 @@ export default {
       },
     });
   },
-  methods: {
-    async deleteCredential(username) {
-      const res = await this.$helpers.fetch('/api/accounts/credentials', {
-        method: 'DELETE',
-        body: JSON.stringify({ username })
-      });
-      if (!res) return;
-
-      this.$store.commit('setAccount', res.account);
-      if (!res.account) {
-        this.$store.commit('setUsername', null);
-        this.$router.push({ name: 'Login' });
-      }
-      this.$store.commit('alert', {
-        message: `Successfully removed ${username} from your account!`, status: 'success'
-      });
-    }
-  },
 };
 </script>
-
-<style scoped>
-li {
-  list-style-type: none;
-  margin-bottom: 1rem;
-}
-
-button {
-  margin-right: 2rem;
-}
-</style>
